+++ libraries/template-haskell/Language/Haskell/TH/Syntax.hs
@@ -1,12 +1,15 @@
-{-# LANGUAGE CPP, DeriveDataTypeable, PolymorphicComponents,
+{-# LANGUAGE CPP, DeriveDataTypeable,
              DeriveGeneric, FlexibleInstances, DefaultSignatures,
-             ScopedTypeVariables, Rank2Types #-}
+             RankNTypes, RoleAnnotations, ScopedTypeVariables #-}
 
-{-# LANGUAGE RoleAnnotations #-}
 {-# OPTIONS_GHC -fno-warn-inline-rule-shadowing #-}
 
 #if MIN_VERSION_base(4,8,0)
 #define HAS_NATURAL
+#endif
+
+#if MIN_VERSION_base(4,9,0)
+# define HAS_MONADFAIL 1
 #endif
 
 -----------------------------------------------------------------------------
@@ -49,13 +52,21 @@ import Language.Haskell.TH.LanguageExtensions
 import Numeric.Natural
 #endif
 
+#if HAS_MONADFAIL
+import qualified Control.Monad.Fail as Fail
+#endif
+
 -----------------------------------------------------
 --
 --              The Quasi class
 --
 -----------------------------------------------------
 
+#if HAS_MONADFAIL
+class Fail.MonadFail m => Quasi m where
+#else
 class (Applicative m, Monad m) => Quasi m where
+#endif
   qNewName :: String -> m Name
         -- ^ Fresh names
 
@@ -173,8 +184,17 @@ runQ (Q m) = m
 instance Monad Q where
   Q m >>= k  = Q (m >>= \x -> unQ (k x))
   (>>) = (*>)
+
   return     = pure
+
+#if !HAS_MONADFAIL
   fail s     = report True s >> Q (fail "Q monad failure")
+#else
+  fail       = Fail.fail
+
+instance Fail.MonadFail Q where
+  fail s     = report True s >> Q (Fail.fail "Q monad failure")
+#endif
 
 instance Functor Q where
   fmap f (Q x) = Q (fmap f x)